reorg util methods in FlinkRelOptUtil
