+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/nodes/logical/FlinkLogicalAggregate.scala
@@ -20,7 +20,7 @@ package org.apache.flink.table.plan.nodes.logical
 
 import org.apache.flink.table.plan.PartialFinalType
 import org.apache.flink.table.plan.nodes.FlinkConventions
-import org.apache.flink.table.plan.util.FlinkRelOptUtil
+import org.apache.flink.table.plan.util.AggregateUtil
 
 import org.apache.calcite.plan._
 import org.apache.calcite.rel.RelNode
@@ -68,7 +68,7 @@ class FlinkLogicalAggregate(
   }
 
   override def computeSelfCost(planner: RelOptPlanner, mq: RelMetadataQuery): RelOptCost = {
-    if (getGroupSets.size > 1 || FlinkRelOptUtil.getGroupIdExprIndexes(getAggCallList).nonEmpty) {
+    if (getGroupSets.size > 1 || AggregateUtil.getGroupIdExprIndexes(getAggCallList).nonEmpty) {
       planner.getCostFactory.makeInfiniteCost()
     } else {
       val child = this.getInput
@@ -103,7 +103,7 @@ private class FlinkLogicalAggregateBatchConverter
       case _ => true
     }
 
-    val hasAccurateDistinctCall = FlinkRelOptUtil.containsAccurateDistinctCall(agg.getAggCallList)
+    val hasAccurateDistinctCall = AggregateUtil.containsAccurateDistinctCall(agg.getAggCallList)
 
     !hasAccurateDistinctCall && supported
   }
