+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/util/FlinkRelOptUtil.scala
@@ -17,21 +17,11 @@
  */
 package org.apache.flink.table.plan.util
 
-import org.apache.flink.table.api.TableException
-
 import org.apache.calcite.plan.RelOptUtil
 import org.apache.calcite.rel.RelNode
-import org.apache.calcite.rel.core.{AggregateCall, JoinInfo}
-import org.apache.calcite.rex.{RexLiteral, RexNode}
-import org.apache.calcite.sql.{SqlExplainLevel, SqlKind}
-import org.apache.calcite.util.ImmutableIntList
-import org.apache.calcite.util.mapping.IntPair
+import org.apache.calcite.sql.SqlExplainLevel
 
 import java.io.{PrintWriter, StringWriter}
-import java.util
-
-import scala.collection.JavaConversions._
-import scala.collection.mutable
 
 object FlinkRelOptUtil {
 
@@ -70,138 +60,4 @@ object FlinkRelOptUtil {
     sw.toString
   }
 
-
-  /**
-    * Get unique field name based on existed `allFieldNames` collection.
-    * NOTES: the new unique field name will be added to existed `allFieldNames` collection.
-    */
-  def buildUniqueFieldName(
-      allFieldNames: util.Set[String],
-      toAddFieldName: String): String = {
-    var name: String = toAddFieldName
-    var i: Int = 0
-    while (allFieldNames.contains(name)) {
-      name = toAddFieldName + "_" + i
-      i += 1
-    }
-    allFieldNames.add(name)
-    name
-  }
-
-  /**
-    * Returns indices of group functions.
-    */
-  def getGroupIdExprIndexes(aggCalls: Seq[AggregateCall]): Seq[Int] = {
-    aggCalls.zipWithIndex.filter { case (call, _) =>
-      call.getAggregation.getKind match {
-        case SqlKind.GROUP_ID | SqlKind.GROUPING | SqlKind.GROUPING_ID => true
-        case _ => false
-      }
-    }.map { case (_, idx) => idx }
-  }
-
-  /**
-    * Returns whether any of the aggregates are accurate DISTINCT.
-    *
-    * @return Whether any of the aggregates are accurate DISTINCT
-    */
-  def containsAccurateDistinctCall(aggCalls: util.List[AggregateCall]): Boolean = {
-    aggCalls.exists(call => call.isDistinct && !call.isApproximate)
-  }
-
-  /**
-    * Returns whether any of the aggregates are approximate DISTINCT.
-    *
-    * @return Whether any of the aggregates are approximate DISTINCT
-    */
-  def containsApproximateDistinctCall(aggCalls: util.List[AggregateCall]): Boolean = {
-    aggCalls.exists(call => call.isDistinct && call.isApproximate)
-  }
-
-  /**
-    * Returns limit start value (never null).
-    */
-  def getLimitStart(offset: RexNode): Long = if (offset != null) RexLiteral.intValue(offset) else 0L
-
-  /**
-    * Returns limit end value (never null).
-    */
-  def getLimitEnd(offset: RexNode, fetch: RexNode): Long = {
-    if (fetch != null) {
-      getLimitStart(offset) + RexLiteral.intValue(fetch)
-    } else {
-      // TODO return Long.MaxValue when providing FlinkRelMdRowCount on Sort ?
-      Integer.MAX_VALUE
-    }
-  }
-
-  /**
-    * Check and get join left and right keys.
-    */
-  def checkAndGetJoinKeys(
-      keyPairs: List[IntPair],
-      left: RelNode,
-      right: RelNode,
-      allowEmptyKey: Boolean = false): (Array[Int], Array[Int]) = {
-    // get the equality keys
-    val leftKeys = mutable.ArrayBuffer.empty[Int]
-    val rightKeys = mutable.ArrayBuffer.empty[Int]
-    if (keyPairs.isEmpty) {
-      if (allowEmptyKey) {
-        (leftKeys.toArray, rightKeys.toArray)
-      } else {
-        throw new TableException(
-          s"Joins should have at least one equality condition.\n" +
-            s"\tleft: ${left.toString}\n\tright: ${right.toString}\n" +
-            s"please re-check the join statement and make sure there's " +
-            "equality condition for join.")
-      }
-    } else {
-      // at least one equality expression
-      val leftFields = left.getRowType.getFieldList
-      val rightFields = right.getRowType.getFieldList
-
-      keyPairs.foreach { pair =>
-        val leftKeyType = leftFields.get(pair.source).getType.getSqlTypeName
-        val rightKeyType = rightFields.get(pair.target).getType.getSqlTypeName
-
-        // check if keys are compatible
-        if (leftKeyType == rightKeyType) {
-          // add key pair
-          leftKeys += pair.source
-          rightKeys += pair.target
-        } else {
-          throw new TableException(
-            s"Join: Equality join predicate on incompatible types. " +
-              s"\tLeft: ${left.toString}\n\tright: ${right.toString}\n" +
-              "please re-check the join statement.")
-        }
-      }
-      (leftKeys.toArray, rightKeys.toArray)
-    }
-  }
-
-  /**
-    * Creates a [[JoinInfo]] by analyzing a condition.
-    *
-    * <p>NOTES: the functionality of the method is same with [[JoinInfo#of]],
-    * the only difference is that the methods could return `filterNulls`.
-    */
-  def createJoinInfo(
-      left: RelNode,
-      right: RelNode,
-      condition: RexNode,
-      filterNulls: util.List[java.lang.Boolean]): JoinInfo = {
-    val leftKeys = new util.ArrayList[Integer]
-    val rightKeys = new util.ArrayList[Integer]
-    val remaining = RelOptUtil.splitJoinCondition(
-      left, right, condition, leftKeys, rightKeys, filterNulls)
-
-    if (remaining.isAlwaysTrue) {
-      JoinInfo.of(ImmutableIntList.copyOf(leftKeys), ImmutableIntList.copyOf(rightKeys))
-    } else {
-      // TODO create NonEquiJoinInfo directly
-      JoinInfo.of(left, right, condition)
-    }
-  }
 }
