+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/util/JoinUtil.scala
@@ -1,0 +1,108 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.table.plan.util
+
+import org.apache.flink.table.api.TableException
+
+import org.apache.calcite.plan.RelOptUtil
+import org.apache.calcite.rel.RelNode
+import org.apache.calcite.rel.core.{Join, JoinInfo}
+import org.apache.calcite.rex.RexNode
+import org.apache.calcite.util.ImmutableIntList
+import org.apache.calcite.util.mapping.IntPair
+
+import java.util
+
+import scala.collection.mutable
+
+/**
+  * Util for [[Join]]s.
+  */
+object JoinUtil {
+
+  /**
+    * Check and get join left and right keys.
+    */
+  def checkAndGetJoinKeys(
+      keyPairs: List[IntPair],
+      left: RelNode,
+      right: RelNode,
+      allowEmptyKey: Boolean = false): (Array[Int], Array[Int]) = {
+    // get the equality keys
+    val leftKeys = mutable.ArrayBuffer.empty[Int]
+    val rightKeys = mutable.ArrayBuffer.empty[Int]
+    if (keyPairs.isEmpty) {
+      if (allowEmptyKey) {
+        (leftKeys.toArray, rightKeys.toArray)
+      } else {
+        throw new TableException(
+          s"Joins should have at least one equality condition.\n" +
+            s"\tleft: ${left.toString}\n\tright: ${right.toString}\n" +
+            s"please re-check the join statement and make sure there's " +
+            "equality condition for join.")
+      }
+    } else {
+      // at least one equality expression
+      val leftFields = left.getRowType.getFieldList
+      val rightFields = right.getRowType.getFieldList
+
+      keyPairs.foreach { pair =>
+        val leftKeyType = leftFields.get(pair.source).getType.getSqlTypeName
+        val rightKeyType = rightFields.get(pair.target).getType.getSqlTypeName
+
+        // check if keys are compatible
+        if (leftKeyType == rightKeyType) {
+          // add key pair
+          leftKeys += pair.source
+          rightKeys += pair.target
+        } else {
+          throw new TableException(
+            s"Join: Equality join predicate on incompatible types. " +
+              s"\tLeft: ${left.toString}\n\tright: ${right.toString}\n" +
+              "please re-check the join statement.")
+        }
+      }
+      (leftKeys.toArray, rightKeys.toArray)
+    }
+  }
+
+  /**
+    * Creates a [[JoinInfo]] by analyzing a condition.
+    *
+    * <p>NOTES: the functionality of the method is same with [[JoinInfo#of]],
+    * the only difference is that the methods could return `filterNulls`.
+    */
+  def createJoinInfo(
+      left: RelNode,
+      right: RelNode,
+      condition: RexNode,
+      filterNulls: util.List[java.lang.Boolean]): JoinInfo = {
+    val leftKeys = new util.ArrayList[Integer]
+    val rightKeys = new util.ArrayList[Integer]
+    val remaining = RelOptUtil.splitJoinCondition(
+      left, right, condition, leftKeys, rightKeys, filterNulls)
+
+    if (remaining.isAlwaysTrue) {
+      JoinInfo.of(ImmutableIntList.copyOf(leftKeys), ImmutableIntList.copyOf(rightKeys))
+    } else {
+      // TODO create NonEquiJoinInfo directly
+      JoinInfo.of(left, right, condition)
+    }
+  }
+}
